added bible command and welcome message
Bible command uses the Bible API to query a user input phrase and
returns an object. An api token was also used so that will need to be
configured in credentials.json.

The onGuildMemberAdd event was registered. Whenvever a new memeber joins
a message will be sent

Various helper functions and classes were added too, including the
PaginationEmbed class used for creating pages of embeds and util
functions like textToSlash which provide backward compatibility with
text based discord commands and parses it to work with the current
interaction based slash command structure